Add check to that build -c works when there is no target directory.

+++ tools/build.js
@@ -419,15 +419,17 @@ task('Remove generated files.', [], function clean() {
     emptyDir(`${APP_HOME}/lib`);
   }
 
-  var files = fs.readdirSync(TARGET_DIR, {withFileTypes: true});
-  files.forEach(f => {
-    // Don't remove java libs under ./target/lib
-    if ( f.name === 'lib' ) return;
+  if ( fs.existsSync(TARGET_DIR) ) {
+    var files = fs.readdirSync(TARGET_DIR, {withFileTypes: true});
+    files.forEach(f => {
+      // Don't remove java libs under ./target/lib
+      if ( f.name === 'lib' ) return;
 
-    var fn = TARGET_DIR + '/' + f.name;
-    if ( f.isDirectory() ) rmdir(fn);
-    if ( f.isFile()      ) rmfile(fn);
-  });
+      var fn = TARGET_DIR + '/' + f.name;
+      if ( f.isDirectory() ) rmdir(fn);
+      if ( f.isFile()      ) rmfile(fn);
+    });
+  }
 
   // TODO: convert to Node to make Windows compatible
   execSync('rm -f foam-bin*.js');
